Singletons and Audio
Added a simple audio manager and put sound effects on the focus and the UI buttons. Implemented magic chamber as a singleton that actually works in the editor as a single location for the game's focus and well, so I don't have to pass variables around all over the place.